backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
boto3==1.34.69
django-storages==1.14.2
celery==5.3.4
redis==5.0.1
django-filter==23.5
drf-yasg==1.21.7
dj-database-url==2.1.0
django-celery-beat==2.5.0
gunicorn==21.2.0 